// src/api/openapi/auto-generated/models/AnalysisSettingsWithMetadataResponse.ts
export interface AnalysisSettingsWithMetadataResponse {
    /**
     * 
     * @type {string}
     * @memberof AnalysisSettingsWithMetadataResponse
     */
    analysisId?: string;
    /**
     * 
     * @type {AdaptationOptions}
     * @memberof AnalysisSettingsWithMetadataResponse
     */
    adaptationOptions?: AdaptationOptions | null;
    /**
     * 
     * @type {AdaptationAnalysisSettings}
     * @memberof AnalysisSettingsWithMetadataResponse
     */
    analysisSettings?: AdaptationAnalysisSettings | null;
    /**
     * 
     * @type {boolean}
     * @memberof AnalysisSettingsWithMetadataResponse
     */
    settingsConfirmed?: boolean | null;
    /**
     * 
     * @type {AnalysisStatus}
     * @memberof AnalysisSettingsWithMetadataResponse
     */
    status?: AnalysisStatus | null;
    /**
     * 
     * @type {ParameterBounds}
     * @memberof AnalysisSettingsWithMetadataResponse
     */
    discountRate?: ParameterBounds | null;
    /**
     * 
     * @type {ParameterBounds}
     * @memberof AnalysisSettingsWithMetadataResponse
     */
    annualGrowth?: ParameterBounds | null;
}