// src/api/openapi/auto-generated/models/LocationAssetAttributes.ts
export interface LocationAssetAttributes {
    /**
     * Globally-unique location identifier.
     * @type {string}
     * @memberof LocationAssetAttributes
     */
    locationId: string;
    /**
     * Customer-provided location ID. Note that a customer location ID is not globally unique, but will be unique within a portfolio. 
     * @type {number}
     * @memberof LocationAssetAttributes
     */
    customerLocationId: number;
    /**
     * 
     * @type {string}
     * @memberof LocationAssetAttributes
     */
    countryCodeIso2a?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LocationAssetAttributes
     */
    entity?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LocationAssetAttributes
     */
    locationName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LocationAssetAttributes
     */
    occupancyName?: string | null;
    /**
     * 
     * @type {number}
     * @memberof LocationAssetAttributes
     */
    occupancyCode?: number | null;
    /**
     * 
     * @type {string}
     * @memberof LocationAssetAttributes
     */
    occupancyScheme?: string | null;
    /**
     * 
     * @type {string}
     * @memberof LocationAssetAttributes
     */
    admin1Code?: string | null;
    /**
     * 
     * @type {number}
     * @memberof LocationAssetAttributes
     */
    totalValue?: number | null;
    /**
     * 
     * @type {number}
     * @memberof LocationAssetAttributes
     */
    unadaptedAal2025?: number | null;
    /**
     * 
     * @type {number}
     * @memberof LocationAssetAttributes
     */
    unadaptedAal2055?: number | null;
}